Move SPA into src/dashboard, stage.tsx as thin bridge
