+++ projects/packages/podcast/src/dashboard/components/cover-image-control.tsx
